Transfer can´t remove 85 % and more percentage of category tree

+++ src/Model/Category/CategoryFacade.php
@@ -5,6 +5,8 @@ declare(strict_types=1);
 namespace App\Model\Category;
 
 use App\Model\Advert\Advert;
+use App\Model\Category\Transfer\CategoryRemoveCronModule;
+use App\Model\Category\Transfer\Exception\MaximumPercentageOfCategoriesToRemoveLimitExceeded;
 use Shopsys\FrameworkBundle\Component\Domain\Config\DomainConfig;
 use Shopsys\FrameworkBundle\Model\Category\CategoryFacade as BaseCategoryFacade;
 use Shopsys\FrameworkBundle\Model\Product\Product;
@@ -74,9 +76,14 @@ class CategoryFacade extends BaseCategoryFacade
      * @param \Shopsys\FrameworkBundle\Component\Domain\Config\DomainConfig $domainConfig
      * @return \Shopsys\FrameworkBundle\Model\Category\CategoryWithLazyLoadedVisibleChildren[]
      */
-    public function getCategoriesWithLazyLoadedVisibleAndListableChildrenForParent(Category $parentCategory, DomainConfig $domainConfig): array
-    {
-        $categories = $this->categoryRepository->getTranslatedVisibleAndListableSubcategoriesByDomain($parentCategory, $domainConfig);
+    public function getCategoriesWithLazyLoadedVisibleAndListableChildrenForParent(
+        Category $parentCategory,
+        DomainConfig $domainConfig
+    ): array {
+        $categories = $this->categoryRepository->getTranslatedVisibleAndListableSubcategoriesByDomain(
+            $parentCategory,
+            $domainConfig
+        );
 
         $categoriesWithLazyLoadedVisibleAndListableChildren = $this->categoryWithLazyLoadedVisibleChildrenFactory
             ->createCategoriesWithLazyLoadedVisibleAndListableChildren($categories, $domainConfig);
@@ -120,8 +127,11 @@ class CategoryFacade extends BaseCategoryFacade
      * @param string|null $searchText
      * @return \App\Model\Category\Category[]
      */
-    public function getVisibleAndListableByDomainAndSearchText(int $domainId, string $locale, ?string $searchText): array
-    {
+    public function getVisibleAndListableByDomainAndSearchText(
+        int $domainId,
+        string $locale,
+        ?string $searchText
+    ): array {
         $categories = $this->categoryRepository->getVisibleAndListableByDomainIdAndSearchText(
             $domainId,
             $locale,
@@ -147,7 +157,10 @@ class CategoryFacade extends BaseCategoryFacade
      */
     public function findMallCategoryForProduct(Product $product, int $domainId): ?string
     {
-        return $this->categoryRepository->findMallCategoryForProduct($product->isVariant() ? $product->getMainVariant() : $product, $domainId);
+        return $this->categoryRepository->findMallCategoryForProduct(
+            $product->isVariant() ? $product->getMainVariant() : $product,
+            $domainId
+        );
     }
 
     /**
@@ -167,8 +180,11 @@ class CategoryFacade extends BaseCategoryFacade
      * @param string $delimiter
      * @return string
      */
-    public function getCategoriesNamesInPathAsString(Category $destinationCategory, string $locale, string $delimiter = '/'): string
-    {
+    public function getCategoriesNamesInPathAsString(
+        Category $destinationCategory,
+        string $locale,
+        string $delimiter = '/'
+    ): string {
         $categoriesInPath = $this->getCategoriesInPath($destinationCategory);
 
         $categoriesNamesInPath = [];
@@ -264,7 +280,21 @@ class CategoryFacade extends BaseCategoryFacade
      */
     public function removeCategoriesExceptPohodaIds(array $pohodaIds): array
     {
+        $allCategories = $this->categoryRepository->getAll();
         $categories = $this->categoryRepository->getCategoriesExceptPohodaIds($pohodaIds);
+
+        $categoriesToRemovePercentage = (count($categories) / count($allCategories)) * 100;
+        if ($categoriesToRemovePercentage > CategoryRemoveCronModule::MAX_BATCH_CATEGORIES_REMOVE_PERCENT) {
+            throw new MaximumPercentageOfCategoriesToRemoveLimitExceeded(
+                sprintf(
+                    'Trying to remove %s categories, which is %s percent of whole category tree, removing aborted. Maximum is %s percent.',
+                    count($categories),
+                    $categoriesToRemovePercentage,
+                    CategoryRemoveCronModule::MAX_BATCH_CATEGORIES_REMOVE_PERCENT
+                )
+            );
+        }
+
         foreach ($categories as $category) {
             $this->deleteById($category->getId());
         }
